# Backend/src/api/openapi.yaml
openapi: 3.0.0
info:
  title: StudentSocial
  version: "1.0"
  description: Die StudentSocial Website ist eine Single Page App im Frontend und ein REST Webservice im Backend. Dadurch sollen Studenten und andere Objekte in einer Datenbank gespeichert werden.
  contact:
    name: WWI20B4 Gruppe
    email: mai.julian@student.dhbw-karlsruhe.de
    url: "https://www.google.de"
  license:
    name: ABC-Lizenz
    url: "https://www.google.de"
  termsOfService: "https://www.google.de"
servers:
  - url: "http://localhost:3000"
    description: Lokale Entwicklung
paths:
...
  "/student/{id}":
    parameters:
      - schema:
          type: string
        name: id
        in: path
        required: true
    put:
      summary: Student überschreiben
      tags:
        - Student
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
        "404":
          description: Not Found
      operationId: put-student-id
      description: "Daten eines Studenteintags überschreiben. Dazu muss der Vollständige Datensatz an den Server geschickt werden, inklusive der Felder, die nicht geändert werden."
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Student"
    patch:
      summary: Einzelnes Feld ändern
      tags:
        - Student
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
        "404":
          description: Not Found
      operationId: patch-student-id
      description: "Daten eines Studenteintags überschreiben. Bei der PATCH Anfrage müssen nur die Felder mitgegeben werden, die tatsächlich geändert werden sollen"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Student"
    delete:
      summary: Student löschen
      tags:
        - Student
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
      operationId: delete-student-id
      description: Einen Studenteintrag löschen
    get:
      summary: Student lesen
      tags:
        - Student
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
        "404":
          description: Not Found
      operationId: get-student-id
      description: Details zu einem einzelnen Studenten
...
  "/address/{id}":
    parameters:
      - schema:
          type: string
        name: id
        in: path
        required: true
    get:
      summary: Adresse lesen
      tags:
        - Adresse
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Address"
        "404":
          description: Not Found
      operationId: get-address-id
      description: Details zu einer einzelnen Adresse ermitteln.
    put:
      summary: Adresse überschreiben
      tags:
        - Adresse
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Address"
        "404":
          description: Not Found
      operationId: put-address-id
      description: "Daten eines Adressbucheintrags überschrieben. Bei einer PUT-Anfrage muss hierfür der vollständige Datensatz an den Server geschickt werden, inklusive der Felder, die nicht geändert werden sollen. Die ID kann als einziges Feld nicht geändert werden. Sie wird aus der URL übernommen und im Request Body vom Server ignoriert."
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Address"
        description: ""
    patch:
      summary: Einzelnes Feld ändern
      tags:
        - Adresse
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Address"
        "404":
          description: Not Found
      operationId: patch-address-id
      description: "Einzelne Felder eines Adressbucheintrags überschrieben. Bei einer PATCH-Anfrage müssen hierfür nur die Felder an den Server geschickt werden, die tatsächlich geändert werden sollen. Die ID kann als einziges Feld nicht geändert werden. Sie wird aus der URL übernommen und im Request Body vom Server ignoriert."
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Address"
    delete:
      summary: Adresse löschen
      tags:
        - Adresse
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
      operationId: delete-address-id
      description: Adressbucheintrag löschen
...
  "/user/{id}":
    parameters:
      - schema:
          type: string
        name: id
        in: path
        required: true
    get:
      summary: User lesen
      tags:
        - User
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Not Found
      operationId: get-user-id
      description: Einzelnen User erhalten
    put:
      summary: User überschreiben
      tags:
        - User
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Not Found
      operationId: put-user-id
      description: Daten User überschreiben
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"
        description: User Body
    patch:
      summary: Einzelnes Feld überschreiben
      tags:
        - User
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
      operationId: patch-user-id
      description: Einzelnes Feld überschreiben
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"
        description: Einzelne Felder überschreiben
    delete:
      summary: User löschen
      tags:
        - User
      responses:
        "204":
          description: No Content
      operationId: delete-user-id
      description: User löschen
...
components:
  schemas:
    Student:
      title: Student
      type: object
      description: |-
        Eintrag in der Datenbank um eine Übersicht über alle Studenten zu erhalten.
        Dieser Betseht aus Pronomen, Vorname, Nachname, Geburtsdatum, Geschlecht und Studiengang.
        Die ID wird mit der Datenbank generiert.
      properties:
        _links:
          $ref: "#/components/schemas/HATEOAS_Resource"
        _id:
          type: string
        matrikel_nr:
          type: string
        first_name:
          type: string
        last_name:
          type: string
        birthday:
          type: string
        course:
          type: string
        course_id:
          type: string
    User:
      title: User
      type: object
      description: Ein Eintrag in der User-Tabelle um den Login zu authentifizieren.
      properties:
        _links:
          $ref: "#/components/schemas/HATEOAS_Resource"
        _id:
          type: string
        matrikel_nr:
          type: string
        email:
          type: string
        password:
          type: string
...
    Address:
      title: Address
      type: object
      description: "Ein Eintrag im Adressbuch, bestehend aus Vorname, Nachname, Telefonnumer und E-Mailadresse. Die ID wird von der Datenbank generiert und kann nachträglich nicht mehr verändert werden."
      properties:
        _links:
          $ref: "#/components/schemas/HATEOAS_Resource"
        _id:
          type: string
        first_name:
          type: string
        last_name:
          type: string
        phone:
          type: string
        email:
          type: string
...
    HATEOAS_Resource:
      title: HATEOAS_Resource
      type: object
      description: "HATEOAS-Verlinkungen zu einem einzelnen Datensatz / Adressbucheintrag. Definiert, wie der Eintrag gelesen, aktualisiert, gepatcht oder gelöscht werden kann."
      properties:
        read:
          $ref: "#/components/schemas/HATEOAS_Link"
        update:
          $ref: "#/components/schemas/HATEOAS_Link"
        patch:
          $ref: "#/components/schemas/HATEOAS_Link"
        delete:
          $ref: "#/components/schemas/HATEOAS_Link"
    HATEOAS_Link:
      title: HATEOAS_Link
      type: object
      description: "Eine einzelner HATEOAS-Verlinkung innerhalb des Webservices. Definiert die Adresse, URL und Query Parameter zum Aufruf der jeweiligen Aktion."
      properties:
        url:
          type: string
        method:
          type: string
        query_params:
          type: array
          items:
            type: string
...